\name{analyzePackage}
\alias{analyzePackage}
\title{Analyze a package}
\usage{
  analyzePackage(package, temp = TRUE)
}
\arguments{
  \item{package}{Character. Name of the package to analyze}

  \item{temp}{logical. If the package isn't currently
  installed do you want it to be installed temporarily or
  permanently? TRUE gives a temporary install - FALSE
  installs the package permanently.}
}
\description{
  Analyze a packageo on CRAN
}

